Beat Bracket V2

+++ src/components/ProfileDropdown.jsx
@@ -1,10 +1,12 @@
+import { navigate } from "gatsby";
 import React, { useEffect, useState } from "react";
-import { loadRequest } from "../utilities/helpers";
 import guestProfileImage from "../assets/images/guestProfileImage.png";
+import { getUserInfo } from "../utilities/spotify";
+import LoginButton from "./LoginButton";
 
-const ProfileDropdown = ({ loggedIn }) => {
+const ProfileDropdown = ({ loggedIn, noChanges }) => {
   const [showDropdown, setShowDropdown] = useState(false);
-  const [userInfo, setUserInfo] = useState({
+  const [shownUserInfo, setShownUserInfo] = useState({
     display_name: "Guest",
     id: "",
     images: [
@@ -14,32 +16,45 @@ const ProfileDropdown = ({ loggedIn }) => {
     ],
   });
 
-  function signOut() {
-    setShowDropdown(false);
-    sessionStorage.clear();
+  function handleNaviagtionAttempt(path) {
+    if (noChanges(true)) {
+      navigate(path);
+    }
   }
 
-  async function getUserInfo() {
-    const url = "https://api.spotify.com/v1/me";
-    const response = await loadRequest(url);
-    if (!response["error"]) {
-      if (response.images.length == 0) {
-        response.images.push({
-          url: guestProfileImage,
-        });
-      }
-      return response;
+  function signOut() {
+    if (noChanges(true)) {
+      setShowDropdown(false);
+      sessionStorage.clear();
+      localStorage.setItem("rememberMe", false);
+      navigate("/");
     }
+  }
+
+  function navProfile() {
+    setShowDropdown(false);
+    handleNaviagtionAttempt("/my-brackets");
   }
 
   useEffect(() => {
-    async function getInfo() {
-      setUserInfo(await getUserInfo());
-    }
     if (loggedIn) {
-      getInfo();
+      getUserInfo().then((userInfo) => {
+        if (userInfo !== 1) {
+          setShownUserInfo(userInfo);
+        } else {
+          setShownUserInfo({
+            display_name: "Guest",
+            id: "",
+            images: [
+              {
+                url: guestProfileImage,
+              },
+            ],
+          });
+        }
+      });
     } else {
-      setUserInfo({
+      setShownUserInfo({
         display_name: "Guest",
         id: "",
         images: [
@@ -52,76 +67,90 @@ const ProfileDropdown = ({ loggedIn }) => {
   }, [loggedIn]);
 
   return (
-    <div className="inline-block relative">
-      <button
-        type="button"
-        className="flex items-center rounded-lg transition group shrink-0 border-0 hover:bg-transparent px-0"
-        id="menu-button"
-        aria-expanded="true"
-        aria-haspopup="true"
-        data-dropdown-toggle="dropdownNavbar"
-        onClick={() => {
-          if (loggedIn) {
-            setShowDropdown(!showDropdown);
-          }
-        }}
-      >
-        <img
-          className="object-cover w-10 h-10 rounded-full"
-          src={userInfo.images[0].url}
-          alt={userInfo.display_name}
-        />
-        <p className="hidden ml-2 text-left sm:block">
-          <strong className="block text-s font-bold text-white">
-            {userInfo.display_name}
-          </strong>
-          <span className="text-gray-300 text-xs transition sm:block group-hover:text-white">
-            {userInfo.id}
-          </span>
-        </p>
-        <div hidden={!loggedIn}>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            className="w-5 h-5 ml-4 text-gray-300 transition sm:block group-hover:text-white"
-            viewBox="0 0 20 20"
-            fill="currentColor"
-          >
-            <path
-              fillRule="evenodd"
-              d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
-              clipRule="evenodd"
-            />
-          </svg>
+    <div>
+      {!loggedIn ? (
+        <div className="border rounded">
+          <LoginButton />
         </div>
-      </button>
-      <ul
-        id="dropdownNavbar"
-        className="absolute bg-white text-base z-10 list-none divide-y divide-gray-100 rounded shadow w-full cursor-pointer"
-        aria-labelledby="dropdownDefault"
-        hidden={!showDropdown}
-      >
-        {/* <li>
+      ) : (
+        <div className="inline-block relative">
+          <button
+            type="button"
+            className="flex items-center rounded-lg transition group shrink-0 border-0 hover:bg-transparent px-0"
+            id="menu-button"
+            aria-expanded="true"
+            aria-haspopup="true"
+            data-dropdown-toggle="dropdownNavbar"
+            onClick={() => {
+              if (loggedIn) {
+                setShowDropdown(!showDropdown);
+              }
+            }}
+          >
+            <img
+              className="object-cover w-10 h-10 rounded-full"
+              src={shownUserInfo.images[0].url}
+              alt={shownUserInfo.display_name}
+            />
+            <p className="hidden ml-2 text-left sm:block">
+              <strong className="block text-s font-bold text-white">
+                {shownUserInfo.display_name}
+              </strong>
+              <span className="text-gray-300 text-xs transition sm:block group-hover:text-white">
+                {shownUserInfo.id}
+              </span>
+            </p>
+            <div hidden={!loggedIn}>
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                className="w-5 h-5 ml-4 text-gray-300 transition sm:block group-hover:text-white"
+                viewBox="0 0 20 20"
+                fill="currentColor"
+              >
+                <path
+                  fillRule="evenodd"
+                  d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
+                  clipRule="evenodd"
+                />
+              </svg>
+            </div>
+          </button>
+          <ul
+            id="dropdownNavbar"
+            className="absolute right-0 bg-white text-base z-10 list-none divide-y divide-gray-100 rounded shadow w-full cursor-pointer min-w-fit"
+            aria-labelledby="dropdownDefault"
+            hidden={!showDropdown}
+          >
+            {/* <li>
           <button className="py-2 px-4 items-center whitespace-nowrap flex gap-1 group-hover:bg-gray-200 border-0 w-full group">
             My Brackets
           </button>
         </li> */}
-        <li>
-          <button
-            onClick={signOut}
-            className="py-2 px-4 items-center whitespace-nowrap flex gap-1 group-hover:bg-gray-200 border-0 w-full group"
-          >
-            <span>Sign out</span>
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              className="w-4 h-4 text-secondary transition sm:block group-hover:text-secondary"
-              viewBox="0 0 24 24"
-              fill="currentColor"
-            >
-              <path d="M16 10v-5l8 7-8 7v-5h-8v-4h8zm-16-8v20h14v-2h-12v-16h12v-2h-14z" />
-            </svg>
-          </button>
-        </li>
-      </ul>
+            <li>
+              <button
+                onClick={navProfile}
+                className="py-2 px-4 items-center whitespace-nowrap flex gap-1 group-hover:bg-gray-200 border-0 w-full group"
+              >
+                <span>My Brackets</span>
+              </button>
+              <button
+                onClick={signOut}
+                className="py-2 px-4 items-center rounded-t-none whitespace-nowrap flex gap-1 group-hover:bg-gray-200 border-0 w-full group"
+              >
+                <span>Sign out</span>
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  className="w-4 h-4 text-secondary transition sm:block group-hover:text-secondary"
+                  viewBox="0 0 24 24"
+                  fill="currentColor"
+                >
+                  <path d="M16 10v-5l8 7-8 7v-5h-8v-4h8zm-16-8v20h14v-2h-12v-16h12v-2h-14z" />
+                </svg>
+              </button>
+            </li>
+          </ul>
+        </div>
+      )}
     </div>
   );
 };
